finish addi logic
