+++ ceno_zkvm/src/instructions/riscv/arith_imm.rs
@@ -1,3 +1,14 @@
+#[cfg(not(feature = "u16limb_circuit"))]
+mod arith_imm_circuit;
+#[cfg(feature = "u16limb_circuit")]
+mod arith_imm_circuit_v2;
+
+#[cfg(feature = "u16limb_circuit")]
+pub use crate::instructions::riscv::arith_imm::arith_imm_circuit_v2::AddiInstruction;
+
+#[cfg(not(feature = "u16limb_circuit"))]
+pub use crate::instructions::riscv::arith_imm::arith_imm_circuit::AddiInstruction;
+
 use std::marker::PhantomData;
 
 use ceno_emul::StepRecord;
@@ -10,77 +21,8 @@ use crate::{
 
 use super::{RIVInstruction, constants::UInt, i_insn::IInstructionConfig};
 
-pub struct AddiInstruction<E>(PhantomData<E>);
-
 impl<E> RIVInstruction for AddiInstruction<E> {
     const INST_KIND: ceno_emul::InsnKind = ceno_emul::InsnKind::ADDI;
-}
-
-pub struct InstructionConfig<E: ExtensionField> {
-    i_insn: IInstructionConfig<E>,
-
-    rs1_read: UInt<E>,
-    imm: UInt<E>,
-    rd_written: UInt<E>,
-}
-
-impl<E: ExtensionField> Instruction<E> for AddiInstruction<E> {
-    type InstructionConfig = InstructionConfig<E>;
-
-    fn name() -> String {
-        format!("{:?}", Self::INST_KIND)
-    }
-
-    fn construct_circuit(
-        circuit_builder: &mut CircuitBuilder<E>,
-        _params: &ProgramParams,
-    ) -> Result<Self::InstructionConfig, ZKVMError> {
-        let rs1_read = UInt::new_unchecked(|| "rs1_read", circuit_builder)?;
-        let imm = UInt::new(|| "imm", circuit_builder)?;
-        let rd_written = rs1_read.add(|| "rs1_read + imm", circuit_builder, &imm, true)?;
-
-        let i_insn = IInstructionConfig::<E>::construct_circuit(
-            circuit_builder,
-            Self::INST_KIND,
-            imm.value(),
-            rs1_read.register_expr(),
-            rd_written.register_expr(),
-            false,
-        )?;
-
-        Ok(InstructionConfig {
-            i_insn,
-            rs1_read,
-            imm,
-            rd_written,
-        })
-    }
-
-    fn assign_instance(
-        config: &Self::InstructionConfig,
-        instance: &mut [<E as ExtensionField>::BaseField],
-        lk_multiplicity: &mut LkMultiplicity,
-        step: &StepRecord,
-    ) -> Result<(), ZKVMError> {
-        let rs1_read = Value::new_unchecked(step.rs1().unwrap().value);
-        let imm = Value::new(
-            InsnRecord::imm_internal(&step.insn()) as u32,
-            lk_multiplicity,
-        );
-
-        let result = rs1_read.add(&imm, lk_multiplicity, true);
-
-        config.rs1_read.assign_value(instance, rs1_read);
-        config.imm.assign_value(instance, imm);
-
-        config.rd_written.assign_add_outcome(instance, &result);
-
-        config
-            .i_insn
-            .assign_instance(instance, lk_multiplicity, step)?;
-
-        Ok(())
-    }
 }
 
 #[cfg(test)]
