+++ ceno_zkvm/src/instructions/riscv/arith_imm/arith_imm_circuit.rs
@@ -1,0 +1,84 @@
+use crate::{
+    Value,
+    circuit_builder::CircuitBuilder,
+    error::ZKVMError,
+    instructions::{
+        Instruction,
+        riscv::{constants::UInt, i_insn::IInstructionConfig},
+    },
+    structs::ProgramParams,
+    tables::InsnRecord,
+    witness::LkMultiplicity,
+};
+use ceno_emul::StepRecord;
+use ff_ext::ExtensionField;
+use std::marker::PhantomData;
+
+pub struct AddiInstruction<E>(PhantomData<E>);
+
+pub struct InstructionConfig<E: ExtensionField> {
+    i_insn: IInstructionConfig<E>,
+
+    rs1_read: UInt<E>,
+    imm: UInt<E>,
+    rd_written: UInt<E>,
+}
+
+impl<E: ExtensionField> Instruction<E> for AddiInstruction<E> {
+    type InstructionConfig = crate::instructions::riscv::arith_imm::InstructionConfig<E>;
+
+    fn name() -> String {
+        format!("{:?}", Self::INST_KIND)
+    }
+
+    fn construct_circuit(
+        circuit_builder: &mut CircuitBuilder<E>,
+        _params: &ProgramParams,
+    ) -> Result<Self::InstructionConfig, ZKVMError> {
+        let rs1_read = UInt::new_unchecked(|| "rs1_read", circuit_builder)?;
+        let imm = UInt::new(|| "imm", circuit_builder)?;
+        let rd_written = rs1_read.add(|| "rs1_read + imm", circuit_builder, &imm, true)?;
+
+        let i_insn = IInstructionConfig::<E>::construct_circuit(
+            circuit_builder,
+            Self::INST_KIND,
+            imm.value(),
+            rs1_read.register_expr(),
+            rd_written.register_expr(),
+            false,
+        )?;
+
+        Ok(crate::instructions::riscv::arith_imm::InstructionConfig {
+            i_insn,
+            rs1_read,
+            imm,
+            rd_written,
+        })
+    }
+
+    fn assign_instance(
+        config: &Self::InstructionConfig,
+        instance: &mut [<E as ExtensionField>::BaseField],
+        lk_multiplicity: &mut LkMultiplicity,
+        step: &StepRecord,
+    ) -> Result<(), ZKVMError> {
+        let rs1_read = Value::new_unchecked(step.rs1().unwrap().value);
+        let imm = Value::new(
+            InsnRecord::imm_internal(&step.insn()) as u32,
+            lk_multiplicity,
+        );
+
+        let result = rs1_read.add(&imm, lk_multiplicity, true);
+
+        config.rs1_read.assign_value(instance, rs1_read);
+        config.imm.assign_value(instance, imm);
+
+        config.rd_written.assign_add_outcome(instance, &result);
+
+        config
+            .i_insn
+            .assign_instance(instance, lk_multiplicity, step)?;
+
+        Ok(())
+    }
+}
